move partner images, basic check on events
